comment 13-09-2023

+++ app/Http/Livewire/Admin/AdminEditCategoryComponent.php
@@ -3,7 +3,9 @@
 namespace App\Http\Livewire\Admin;
 
 use App\Models\Category;
+use App\Models\Subcategory;
 use Carbon\Carbon;
+use Illuminate\Support\Str;
 use Livewire\Component;
 use Livewire\WithFileUploads;
 
@@ -14,26 +16,44 @@ class AdminEditCategoryComponent extends Component
 
     public $name;
     public $slug;
-    public $top_category;
-    public $popular_category;
     public $image;
-    public $status;
     public $category_id;
-
+    public $category_slug;
     public $newimage;
 
 
-    public function mount($category_id)
+    public $scategory_id;
+    public $scategory_slug;
+
+
+
+
+
+
+    public function generateSlug(){
+        $this->slug=Str::slug($this->name);
+    }
+
+    public function mount($category_slug,$scategory_slug=null)
     {
 
-        $category = Category::find($category_id);
-        $this->name = $category->name;
-        $this->slug = $category->slug;
-        $this->top_category = $category->top_category;
-        $this->popular_category = $category->popular_category;
-        $this->image = $category->image;
-        $this->status = $category->status;
-        $this->category_id = $category->id;
+        if($scategory_slug){
+            $scategory=Subcategory::where('slug',$this->scategory_slug)->first();
+
+            $this->scategory_id=$scategory->id;
+            $this->name=$scategory->name;
+            $this->slug=$scategory->slug;
+            $this->category_id=$scategory->category_id;
+
+        }else{
+            //$this->category_slug=$category_slug;
+            $category = Category::where('slug',$this->category_slug)->first();
+            $this->name = $category->name;
+            $this->slug = $category->slug;
+            $this->image = $category->image;
+            $this->category_id = $category->id;
+        }
+
     }
 
     public function updateCategory()
@@ -42,34 +62,37 @@ class AdminEditCategoryComponent extends Component
         $this->validate([
             'name' => 'required',
             'slug' => 'required',
-            'top_category' => 'required',
-            'popular_category' => 'required',
-            'status' => 'required',
         ]);
 
-        $category = Category::find($this->category_id);
-        $category->name = $this->name;
-        $category->slug = $this->slug;
-        $category->top_category = $this->top_category;
-        $category->popular_category = $this->popular_category;
-        $category->status = $this->status;
-        if ($this->newimage) {
-            unlink('frontend/assets/images/category/' . $category->image);
-            $imageName = Carbon::now()->timestamp . '.' . $this->newimage->extension();
-            $this->newimage->storeAs('category', $imageName);
-            $category->image = $imageName;
+
+        if($this->scategory_id){
+            $scategory=Subcategory::find($this->scategory_id);
+            $scategory->name=$this->name;
+            $scategory->slug=$this->slug;
+            $scategory->category_id=$this->category_id;
+            $scategory->save();
+        }else{
+            $category = Category::find($this->category_id);
+            $category->name = $this->name;
+            $category->slug = $this->slug;
+            if ($this->newimage) {
+                unlink('frontend/assets/images/category/' . $category->image);
+                $imageName = Carbon::now()->timestamp . '.' . $this->newimage->extension();
+                $this->newimage->storeAs('category', $imageName);
+                $category->image = $imageName;
+            }
+
+            $category->save();
         }
 
-        $category->save();
-
         session()->flash('message', 'Category has been Updated successfully');
-
 
     }
 
 
     public function render()
     {
-        return view('livewire.admin.admin-edit-category-component');
+        $categories=Category::all();
+        return view('livewire.admin.admin-edit-category-component',['categories'=>$categories]);
     }
 }
